move partiticpation in external roles to dedicated folder
